bug fixs

+++ backend/src/main/kotlin/com/burnoutstopper/burnouttest/repository/AnswerRepository.kt
@@ -1,9 +1,9 @@
 package com.burnoutstopper.burnouttest.repository
 
-import Answer
+import com.burnoutstopper.burnouttest.model.Answer
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface AnswerRepository : JpaRepository<Answer, Int> {
+interface AnswerRepository : JpaRepository<Answer, Long> {
 
     fun findByRespondentId(respondentId: Int): Answer
 }
